added auth process to the java plugin
It now checks whether it can establish a connection to the MCConnect server and whether the provided license key is valid. If so the plugin continues to initialize and if not the plugin disables itself.

Note: This isnt required to protect the service for unlicensed access. This whole process could be skipped by source code modifications or direct api calls but the server wouldnt accept any commands without proper auth.